rename client_api to api
